pyproject.toml
[project]
name = "qa-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13, <3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.3",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-dotenv>=0.5.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
